Modify permutate methods to return 'impl Iterator' instead of trait object
Also remove many of the intermediate vecs used in the various methods. I
don't know if this is any faster but it feels more idiomatic, and should
reduce memory usage when scanning lots of domains, since intermediate
allocations should be reduced.

This is obviously a breaking change since it changes the public API
considerably.