fix(permutation): filter permutations by valid tld
During `Domain` initialization we are validating whether
or not the fqdn passed is a valid domain. Prior to this
patch it would also accept structurally valid domains,
which have invalid tlds (e.g., ęxamplę.čom).

This patch adds a bst search on the `TLDS` constant that
we build at compile-time to check that it's a valid tld
as well.